fix: use aws-sdk-js resource waiter `delay` config
Until aws-sdk-js version 2.31.0 was recently released, the resource
waiter that we use for table creation and deletion wasn't configurable.
The (quite unreasonable high) default is `20`, which means that the
waiter waits 20 seconds until checking if the resource exists, or
doesn't exist anymore, respectively.

Now that aws-sdk-js supports configuring the waiter delay, we set it to
`1`.